+++ src/main/java/me/mni_js/springboot_blog/repository/BlogRepository.java
@@ -3,5 +3,5 @@ package me.mni_js.springboot_blog.repository;
 import me.mni_js.springboot_blog.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogRepository extends JpaRepository<Article, Integer> {
+public interface BlogRepository extends JpaRepository<Article, Long> {
 }
